Include storyboards inside .bundles correctly.
This changes the storyboard linking behavior so that it links each bundle's storyboards independently. This implies that an app can't have one of its storyboards link to a storyboard in one of its .bundle directories. This is fine, though, because Xcode already doesn't make it possible to create a .bundle and have its storyboards compiled -- you just end up with the .storyboard copied directly into the bundle, which is useless. We're already going above and beyond what Xcode supports by having storyboards in bundles be compiled in the first place, so this "missing feature" is something that you wouldn't be able to construct in the UI to begin with.

Fixes #5.

PiperOrigin-RevId: 153739673