Don't pass "deps" to the bundler.
Instead, use apple_binary as a choke point for dependencies, even when a target does not need a custom binary (because it uses a stub instead). In the latter case, the binary artifact is never linked (because it is never used as an input), but its providers are still used to obtain other resources.

BUG=37634873
PiperOrigin-RevId: 154225787