Add "resources", "structured_resources" attrs to swift_library.
This *only* works with the Skylark-based bundling rules; the attributes are ignored if a native bundlng rule depends on them.

The "resources" attribute takes the place of "asset_catalogs", "datamodels", "resources", "storyboards", "strings", and "xibs" that are on similar rules like "objc_library". All the specific resource types can be included in "resources" instead and they are processed according to their names.

"structured_resources" has the same meaning as before.

"bundles" are not yet supported; they will be in the near future.

RELNOTES: None.
PiperOrigin-RevId: 154857955