Add "manual" tag to implicit apple_binary targets.
This will prevent builds of :all or /... from trying to build the binary as an isolated target. In some cases doing so would cause failures (e.g., watchOS applications) because the linked binary will error out since it has no srcs but the target must still act as a choke point to correctly pass the platform down to its dependencies.

RELNOTES: None.
PiperOrigin-RevId: 155880519