Refactor the resource processing logic in the bundler.
This fixes an issue where multiple swift_libraries could not propagate asset catalogs because each dependency would try to produce its own Assets.car file in the same location in the bundle. Unfortunately, this required a fairly large refactoring of the processing loop to ensure that resources that do not depend on the Swift module name as an input argument are all processed in a single pass.

Fixes #67.

RELNOTES: None.
PiperOrigin-RevId: 161540238