Decouple entitlements and provisioning profile handling.
This provides better support for macOS, where provisioning profiles are more rare but you may still want entitlements to be set for things like the app sandbox. Now, if a provisioning profile is not set, we don't extract entitlements from it and we don't substitute placeholders using its values, but we still use whatever entitlements the user has provided, if any.

PiperOrigin-RevId: 163247073