Only use -swift-version on Xcode 9
Swift compiler included with Xcode up to 8.3.3 has a bug where swift
language version checks do not work properly with "-swift-version 3"
specified.

For example, swift(>=3.1) does not pass when -swift-version 3 is set,
but does otherwise.

This has been fixed since Xcode 9, and so we bump the safeguard in
swift_library to that version.

PiperOrigin-RevId: 163355745