Fix bug affecting module name at root of workspace
If an Objective-C or Swift target is defined at the root of a workspace, swift_module_name was creating a module name with a `_` prefix. For example, a target named "MyComponent" would be given the name "_MyComponent". This meant that targets depending on that module would have to import the component as _MyComponent rather than the expected MyComponent.

This change fixes that bug by checking whether there is a non-empty package prefix string when generating the module name. If there is no prefix then we simply return the target name.

Closes #109