Dedup `infoplists` when deduping resources between app and frameworks.
The info.plists that will be merged also have to go through the dedup
filtering, otherwise a resource bundle could have all the other resources
removed but still end up creating the .bundle directory and landing the
merged Info.plist into it.

Tweak the test to include and info.plist to show problem and confirm it is
fixed.

RELNOTES: None.
PiperOrigin-RevId: 172746333