Improve plist variable substitutions.
When Xcode does variable substitutions in Info.plist it replaces unknown
keys with the empty string. The long term goes is trap them as errors since
sending them through or just eating them is likely to allow an issue to slip
through.

As a start, throw errors for malformed variable references and print out
a warning for any substitution reference that doesn't map to anything.

RELNOTES: None.
PiperOrigin-RevId: 173113192