Don't create PkgInfo for mac_bundle uses.
Testing with Xcode 9.0.1, it does not create a PkgInfo file for a kext
or spotlight plugin, nor for a resource only bundle. This matches what
would be expected since none of those would have had a old four char
signature to "own" files with, so there shouldn't be a need to capture
that sorta info.

RELNOTES: None.
PiperOrigin-RevId: 173721331