Require CFBundleShortVersionString and CFBundleVersion values.
For most bundle types, require CFBundleShortVersionString and CFBundleVersion
pairs be set per Apple's rules.

This also removes the support for defaulting the versions to 1.0 when nothing
is set.

Future work will start for validate the values, but for now, just require them
to be there and not empty/zero.

RELNOTES: Bundle version numbers are required for most apple rule types.
PiperOrigin-RevId: 176552594