Codesign bug was fixed.
When compiling iOS application for iphoneos with flags
`--ios_multi_cpus=armv7,arm64`, all of compiling have been succeeded.
In the phase of codesign, bazel cannot find the credential file,
`.mobileprovision`.
This seems to be caused by no copy of the credential in the sandbox
building the source codes.

In details, the sandbox would copy the owned credential as
`embedded.mobileprovision` into sandbox directory.
Therefore, this code used `find . -name 'embedded.mobileprovision'` as
a provisioning profile instead of a specified path in `BUILD` rules,
which cannot be found in sandbox.