Refactor the UIDeviceFamily key setting.
Instead of returning the value to the plist code to then have it know about
some platformisms of the family to number lookup, just expose a
platform_support method to do all the work and return the array (or None)
so the plist code minimally acts on what it gets back.

There are already tests on the rules to ensure these values are set correctly.

PiperOrigin-RevId: 189791569