Remove providers.find_one().
Was only used in one place, and the definition is a little confusing
because the docs sorta implied it worked on a single instance or list
but always returned a list, but the one usage says it never returned
a list.  So just remove it rather than having it poorly defined.

PiperOrigin-RevId: 195141325