Fix -add_ast_path link option.
Because this option was being passed as a single string containing spaces, it was never actually being used by clang; this was only discovered by the use of -Wunused-command-line-argument. Unfortunately since multiple `-Xlinker` flags will get deduped if passed separately in a depset, we need to switch to the single-string `-Wl,` form.

RELNOTES: Fix -add_ast_path link option in swift_library.
PiperOrigin-RevId: 197158420