Add support for using import in the pyton tools.
While I haven't seen it documented, the fact that the WORKSPACE name exposed in
the .runfiles can be used to help scope python imports to ensure we are getting
our modules at runtime. We already use this directory structure for our shell
script tools:
  https://github.com/bazelbuild/rules_apple/blob/master/tools/actoolwrapper/actoolwrapper.sh#L26

The "catch" is while a py_binary when invoked by an action always see a
directory structure with the needed __init__.py files to make things modules,
py_test does not create them at the top most layer. So we add the top level one
and the dependency on it to ensure it is always there for the tests.

If/when bazel's py_test gets fixed and we are ok depending on that bazel version
for our tests, this change can be reverted.

PiperOrigin-RevId: 199176152