Remove the support for halting CoreSimulatorService.
It appears there has been a typo in this for a long time now and it
should have been com\.apple\.CoreSimulator\.CoreSimulatorService for
the pattern. That means even if someone were enabling this, it was
doing nothing.

Doing some local testing using the *_resource_* tests for our flake
failure rate doesn't seem to be any different with/without this working
correctly. So just drop the support to simplify the scripts.

PiperOrigin-RevId: 199463961