- Stop propagating toolchain linkopts in rules_apple's swift_library. - Update binary support to add the Swift linkopts to the binary for all platforms.
This unifies the way Swift linkopts are handled in preparation for the migration to rules_swift's swift_library. It has the unfortunate side effect of introducing a new swift-linkopts-propagating target in the build graph for all platforms now instead of just macOS, but we can eventually remove this when we update apple_common.link_multi_arch_binary to take extra linkopts.

This change also moves the test rules' binaries to call binary_support.create_linked_binary, which unifies all the places where the Apple rules create binaries. This will make it easier to update all the call sites when we migrate to the apple_common API above.

PiperOrigin-RevId: 205674615