Hook the experimental bundling implementation into the main implementation. This allows developers to test the new bundling logic by specifying the --define=apple.experimental.bundling=1 flag.
This new bundling implementation uses a different schema for intermediate files, so when switching to the experimental bundling logic, you will see that the resource processing actions will run again.

The new bundling implementation should produce the exact same outputs as the previous implementation.

RELNOTES: A new bundling logic is now available when building Apple bundles. In order to test it, add --define=apple.experimental.bundling=1 to your build flags. This logic will become the default in the future.
PiperOrigin-RevId: 210609101