Make clangrttool run under Python 3.
The only remaining fix was to return strings (py2 unicode) instead of bytes for
a header processing function.

NOTE: In order to run under Python 3, the macholib library must be availble.
It is installed by default for OS X's system Python 2, but is not available
by default for Python 3. macholib is not added as a Bazel dependency to
clangrttool because of low Python 3 usage with the Apple rules. If you're running
Python 3, either install macholib for the interpreter Bazel uses, or file
an issue to have the clangrttool BUILD rules updated appropriately.
PiperOrigin-RevId: 212296587