Allow plists to not be validated for missing variables
This adds a new control key for plisttool to not validate that all keys
have been replaced after processing. This is useful if you have a custom
rule that replaces some plist keys, but leaves the defaults around to be
processed later. This allows users to implement a simple rule to work
around this limitation https://github.com/bazelbuild/rules_apple/issues/226