Add support for LLVM coverage collection to Apple test rules.
Bazel's coverage collection script will be augmented to honor the `TEST_BINARIES_FOR_LLVM_COV` environment variable, which is a semicolon-delimited list of binary paths (runfiles-relative) that must be passed to `llvm-cov` so that coverage data can be mapped back to the binaries. This supports multiple binaries because we need to be able to cover test bundles and test hosts in the same run.

RELNOTES: None.
PiperOrigin-RevId: 222073790