Remove the providers helper and update the code.
This used to be a little more complex as provides changed between styles,
now it don't really add value.

One usage was only using the first result, so not using the helper and
returning early will be performance improvement in most cases.

The other checks for multiple and fails, so it could fail fast, but
the common case will be to only find one, so a simple loop isn't
harmful. And doing the loop as one line is less confusing since there
is no indirection.

PiperOrigin-RevId: 226030472