Drop use of private_fallback on platform_type.
platform_type is used by apple_binary, so it has to be public at the
moment (lots of comments in the code already for this). So make the one
private usage match the others, and remove the need for private fallback
until things can move to the linking api instead.

RELNOTES: None.
PiperOrigin-RevId: 227566643