Add codesigntool to handle multiple signing identities
In order to support profiles with multiple identities, the shell command that extracts the signing identity has been replaced by a python tool that wraps `codesign`, and is able to handle `.mobileprovision` files with multiple signing identities.

This is a continuation #178 to resolve the CLA issues.

I also rebased on the latest master.

Closes #232.

RELNOTES: None
PiperOrigin-RevId: 227855800