Add support for tree artifact outputs under a flag.
This change adds support for creating tree artifact outputs when passing the --define=apple.experimental.tree_artifact_outputs=1 flag.

As part of this work, refactor the embedded_bundle partial to deduplicate processing code. This allows us to have 1 decision point for processing the embedded bundles when the tree artifact outputs are enabled.

This flag enables tree artifacts for the requested target _and_ all dependency bundles as well.

This flag, as the name implies, should be considered experimental and results may be volatile.

RELNOTES: None
PiperOrigin-RevId: 228572596