Stop using `type()` and move to skylib's types.
This flips the then/else clause also, in checking if something was a
struct, it was also dependent on skylib's `partial` implementation
details. If starlark ever gets some other constructs to allows it to
move off the `struct` usage, this could would have had to be updated
anyways.

RELNOTES: None.
PiperOrigin-RevId: 232341811