Inline clang_support into the clang partial.
Now that the partial is used, there aren't multiple callers, so having
all the logic in one place is easier to follow.

Drop one "helper" method that just made the action, it was wrapping one
call, so it seemed clearer to just inline that change since it was about
the same size.

RELNOTES: None.
PiperOrigin-RevId: 233601499