Simplify the running of buildifier on CI.
As of 0.22.0 there are enough improvements in outputs and exit codes that
it seems to make more sense to no longer try to pretty the results and
just rely on a run and exit status to add a footer telling folks how
to fix up things if something is found.

RELNOTES: None.
PiperOrigin-RevId: 234603600