Fix conflict test environment issues
Previously the test_env was being replaced in the test runner template
by the runner itself. At this point any value set for the `env` attr in
the test rule was not realized. Now we instead do the replacement in the
test rule itself, including the environment added by the runner.

RELNOTES: Support passing env in test rules to the test runner