Add /usr/lib/swift to the LC_RPATH list on Xcode 10.2.
When building with Xcode 10.2, if the target uses Swift, the /usr/lib/swift rpath will be injected at the beginning of the rpath list. This will ensure that the system Swift dylibs are prioritized.

When building with Xcode < 10.2, there's no need to inject /usr/lib/swift, since the OS will detect that the application was built with an old version of Xcode and is allowed to run with the bundled dylibs.

RELNOTES: None
PiperOrigin-RevId: 236687885