Resource aspect cleanup for merge_zip provider field.
Only add the merge_zip resources if it's non-empty, otherwise it will propagate an empty list.

Also return early from the resources partial if there are no resource providers to process. This is the case for ios_static_framework, which can be perfectly valid without any kind of resources.

RELNOTES: None
PiperOrigin-RevId: 241735189