Add utf8 decode to better support py2 and py3.
This is the same as what was done in execute.py and it was done in two
of the three cases here, so this makes things more consistent.

There is one place where execute could be used directly to invoke/filter,
but it isn't clear if that is as worth it since the other process helper
would still be needed to process the output vs. just relay it.

RELNOTES: None.
PiperOrigin-RevId: 252469957