Fix framework imports with Swift modules
With bazel 0.27.0 and the flip of `incompatible_objc_framework_cleanup`
https://github.com/bazelbuild/bazel/issues/7944 framework imports of
Swift frameworks broke because the modules were included and processed
as module map files. This change makes them available in the build
but without processing them. They are handled automatically by clang /
swift since the framework is included using `-F`