Move to shared environment.plist generation
This introduces a new rule that bundling rules depend on to generate the
environment plist a single time. This file is required when uploading
builds to Apple, and doesn't change if you build many similar targets.
This can greatly decrease build times if you have a lot of targets that
require this.