Adding a new 'resources' field to all Apple rules that support bundling. The resources attribute is used for adding files to the bundles resouce list that are not otherise associated with any objc_library or swift_library. The resources will be made available in the resulting binaries resources folder. For example, a `macos_application` would include the resources in the `<name>.app/Contents/Resources` folder and an `ios_application` would include the resources in the `<name>.app/Resources` folder.
RELNOTES: Adding a new 'resources' field to all Apple rules that support bundling. The resources attribute is used for adding files to the bundles resouce list that are not otherise associated with any objc_library or swift_library. The resources will be made available in the resulting binaries resources folder. For example, a `macos_application` would include the resources in the `<name>.app/Contents/Resources` folder and an `ios_application` would include the resources in the `<name>.app/Resources` folder.
PiperOrigin-RevId: 262442773