Ensuring the absolute path is used for actool actions.
This prevents tool failures if an asset catalog is located at the root of a workspace as the build process creates symbolic links to the original source folder during the build. actool appears to not follow symbolic links.

RELNOTES:
Ensuring the absolute path is used for actool actions.

This prevents tool failures if an asset catalog is located at the root of a workspace as the build process creates symbolic links to the original source folder during the build. actool appears to not follow symbolic links.

PiperOrigin-RevId: 262953474