Fix codesigningtool.py py3 compatibility.
In recent versions of plistlib, binary data entries are returned as instances of the built-in bytes class, and plistlib.Data is deprecated.
Since this script was expecting a plistlib.Data, it would fail with the error "AttributeError: 'bytes' object has no attribute 'data'".
This change makes it compatible with both new and old versions of plistlib.