Make codesigningtool compatible with Bazel 0.27+.
Prior to Bazel 0.27, when no python_version attribute was specified in a
py_binary target, Python 2 was used as the default interpreter version.
After 0.27, the default interpreter version was switched to Python 3.
Although the binary is marked as source compatible with both Python 2
and Python 3, gRPC has seen failures in its CI system from this binary
while attempting the Bazel upgrade.

This change ensures that the binary will run under an interpreter with
which it is actually compatible.