Pass is_test to swift_runtime_linkopts
rules_swift uses this flag to determine if extra linker options specific
to Swift should be propagated. Previously this worked just because there
is some duplication in which flags are passed. With Xcode 11 a new
directory has been added that we only want to pass to the linker for
tests.