Make dynamic framework slicer more robust for some inputs.
This change replaces the `awk` command that was used to flatten the `lipo` outputs, since it wasn't actually doing the right thing in all cases. It also corrects a bug where the wrong variable was being used at the end to construct the list of slices to pass to the final `lipo` command.

RELNOTES: None.
PiperOrigin-RevId: 278438090