Suppress function-docstring-args in specific places.
If a function is over a computed complexity, buildifier requires a comment, if
the args aren't documented, that then triggers a warning for getting the args
documented. For these cases were we don't really care about the args because
things are pass thru to the real rule, just suppress the warning. Then then
enabled the real arg validation warnings so in the cases we do document args,
they will be correct/complete.

RELNOTES: None
PiperOrigin-RevId: 295140570