Changes to resource processing for more efficient memory usage.
Switching from dictionaries to depsets allows for more references instead of redundant copies, as revealed by an earlier change. Added a note about why we do this in upstream work to collect these resources.

Also avoiding the processing of "unprocessed" resources at the upstream rule level, as it maps a noop to a noop and wastes CPU cycles for no benefit in the short term. The hope was to address a deduplication case that exists as long as unprocessed resources are still processed at the top level rule. Ideally, we won't have to worry about wasted efforts with deduplication vs top level rule resources in the long term.

RELNOTES: None.
PiperOrigin-RevId: 321717336