Support custom executable name for rules that create executables. (#876)
* Support custom executable name for rules that create executables.

This adds a new attribute `executable_name` to rules that create
executables that allows specifying custom executable name in the final
bundle. This is an optional attribute that, if not specified, will get the
value of `bundle_name` or `name` in this order.

This new attribute can be useful, for example, when building a target
with multiple configurations, that it allows you to specify different
executable names for different configurations, while still keeping the
archive bundle name and dSYM bundle name unchanged.

* Add a custom `bundle_name` to `app_with_custom_executable_name` test app
to verify the order of operations prefers the custom `executable_name`.