Preserve the codesign load commands in the output binary when stripping bitcode from Swift standard libraries.
This is needed because the Swift standard libaries bundled in the
SwiftSupport directory in the .ipa file aren't resigned by the
developer's certificate --- they're originally signed by Apple and we're
not supposed to change it.

This change adds the `-keep_cs` flag to the `bitcode_strip` wrapper, and
only run it when building with Xcode 11.4 or above. This preserves the
codesign load commands when stripping bitcode from the Swift standard
libraries, thus ensures those `.dylib`s keep the original code signature
from Apple if we don't resign them.

Note that Swift standard libraries bundled under
`Payload/YourApp.app/Frameworks` are (re)signed with the developer's
certificate --- this change does not affect them.