Do not modify Swift standard libaries bundled in SwiftSupport
This is needed because the Swift standard libraries bundled in the
SwiftSupport directory in the .ipa file shouldn't be resigned by the
developer's certificate---they're already signed by Apple and we're
not supposed to change it.

With this change, when building without bitcode, it will register an
additional action for copying Swift standard libaries for bundling for
SwiftSupport. This action does not touch those dylibs regardless of the
current bitcode mode.

Note that Swift standard libraries bundled under
`Payload/YourApp.app/Frameworks` are still (re)signed with the
developer's certificate---this change does not affect them.