Add a feature to include imported frameworks' module maps in Obj-C compilation inputs
With the flip of `--incompatible_objc_framework_cleanup`, the
`objc_library` implementation in Bazel no longer passes module maps as
inputs to the compile actions, so that `@import` statements for
user-provided framework no longer work in a sandbox. This trap door
allows users to continue using `@import` statements for imported
framework by adding module map to header_imports so that they are
included in Obj-C compilation but they aren't processed in any way.