Ignore .DS_Store files in the validation of single xcassets type
`.DS_Store` doesn't represent an xcassets type -- those files don't
prevent the asset catalog compile action to execute. They can be
generated and accidentally included if the user open the `.xcassets`
directory in the Finder. This ignores all the `.DS_Store` files in the
validation logic, to avoid failing the build unnecessarily.