Ensure that output groups returned from `apple_common.link_multi_arch_binary` are propagated by the bundling rules.
This is important in case any of these actions are validating actions (which use a specially-named output group); we need to make sure those get merged with any output groups that the bundling rules create.

This CL also adds some logic to make sure output groups with the same name get their file sets merged, instead of simply taking the last occurrence of the name, so that output groups aren't silently lost.

PiperOrigin-RevId: 335076331