Fix simulator running with tree artifacts
The tree artifacts bundler correctly sets the permissions of the app and
its nested directories to be writable but then bazel resets them to not
be writable. This doesn't apply for the ipa codepath since the
permissions are retained in the zip. This fixes this by copying the
output .app from bazel into a tempdir and chmoding all the nested
directories. Without this the `simctl install` fails.