Improve reliability of `bazel run` for `ios_application` when launching Simulator app
`bazel run` on an `ios_application` when `Simulator.app` is not yet
running can currently fail with an error like:

```
ERROR ['osascript', '-e', 'tell application "Simulator" to activate'] exited with error code 1
```

Subsequent runs succeed.

The root cause is that launching `Simulator.app` via
`subprocess.Popen()` and then immediately trying to send it an Apple
Event with `osascript` to bring it to the foreground is racy: the
newly-launched `Simulator` process may or may not have yet connected
to the Apple Events system.

This resolves the issue by using the Apple command-line utility `open
-a /path/to/Simulator.app` to launch (if needed) and then bring
`Simulator.app` to the foreground in one command.

RELNOTES: bazel run for ios_application targets now works more reliably when Simulator.app is not yet running
PiperOrigin-RevId: 350411066