Validate for all case-variants of invalid top-level directories in bundles (#1029)
Previously, we only checked for the non-existent top level `Resources`
folder, but not other variants like `resources`. If the developer uses
any of those names, the validation would pass and then fail in the
codesign action.

Since the build may be running on a case-insensitive file-system, which
is the default file-system for macOS, the validation should check for
all the case variants of invalid directory names.