Support for Alternate Icons (#1032)
Because alternate icons unfortunately can't be packaged as `appiconset`, alternate icons must be directories in the form of `<alticon_id>.alticon`, such as `AppIcon-Dark.alticon` if `AppIcon-Dark` is the id. The `PNGs` inside can be named freely.

For instance, with an alticon id as `Dark`:
```
MyApp/Resources/Dark.alticon/
├── AppIcon-Dark@2x.png
├── AppIcon-Dark@2x~ipad.png
├── AppIcon-Dark@2x~ipadpro.png
└── AppIcon-Dark@3x.png
```

They can be added via the `ios_application.alternate_icons` attribute. The icons are processed like any other `png` file and will be embedded as a top level resource.

Some implementation notes:
Since there is no `string_keyed_label_list_dict`, the `alticon` trick is a way to both group the icons together and explicit the icon id.

Also, since all intermediate plists are merged at the top level, `CFBundleIcons.CFBundleAlternateIcons` needs to be generated in the same partial as `CFBundleIcons.CFBundlePrimaryIcon`. This is done by detecting if there are alternate icons, and if so, reinject the `plist` into `alticonstool`, which will modify it accordingly.

Refs #895 
Resolves #895 

Signed-off-by: Steeve Morin <steeve@zen.ly>