Introduce minimum_deployment_os_version
minimum_deployment_os_version allows to decorelate the bundle MinimumOSVersion
from the minimum_os_version which also controls the compilation flags.

This allows to build most modules only once (provided new features are
guarded, also with weak frameworks). However, when doing trying to
install such bundle on iOS < 14, it will fail with:

    This app contains an app extension that specifies an extension point
    identifier that is not supported on this version of iOS for the value of
    the NSExtensionPointIdentifier key in its Info.plist.

This is because the NSExtensionPointIdentifier=com.apple.widgetkit-extension is
not recognized.

It is possible, however, to still compile with one version, but declare a
different MinimumOSVersion in the bundle's plist.

Signed-off-by: Steeve Morin <steeve@zen.ly>