Introduce first stardoc API docgen
This replaces some hand-maintained documentation with the output of the documentation generator.
I merged the contents by porting some of the hand-edits back to the API docstrings.

This gives us a pattern to follow to incrementally replace the docs with generated equivalents.

Note that the documentation has to be organized into .md files that are 1:1 with .bzl files we generate from (a starlark 'module')

See #867