Stardoc rules to generate API docs for watchos
Changes the migration pattern to match existing filenames. This means we don't need so much link redirection, and causes less churn on the git history.
Adds a pattern of [rules].doc.bzl which exports the documentable rules, rather than the macros that users load.
(We use this pattern in rules_nodejs as well)