Create partial to generate codesigning dossiers in a new `dossiers` output group.
Creates a partial to create codesigning dossiers for use in 'one step' signing processes in future. Only enables for iOS and watchOS rules. Enabled only if you request `dossiers` output group, will follow up with enabling for more rules.

Relatedly, adds an option to infer the codesigning identity to the dossier signing script, to enable using the provisioning profile and certificates present in the keychain to determine certificate used.

PiperOrigin-RevId: 374435623