Add additional_linker_inputs to linking attributes
This is to support linkopts with input files, and with $(location)
expansion.  Much of the support is done by native code that
automatically does the right thing if the additional_linker_inputs
attribute is on the rule.  This is similar to how the linkopts
attribute is supported now.

PiperOrigin-RevId: 375142378
(cherry picked from commit 48d3dcb6955c21cb77134975646281ce59fee64a)