Add `swift_clang_module_aspect` to `apple_{dynamic,static}_framework_import.deps`.
Without this, when a framework target depends on another framework target and they
are dependencies of a Swift target, the parent framework does not see the
`SwiftInfo` provider created by the aspect from the child target's
`swift_interop_info`. It's not clear to me why this happens, but there must be
subtle details about the order aspects are applied and when their providers are
merged with the target's main providers -- I think it has to do with the fact
that `swift_clang_module_aspect` doesn't advertise that it provides `SwiftInfo`
(it can't, because it doesn't always).

Directly applying the aspect to the `deps` attribute seems to work though.

PiperOrigin-RevId: 376191209