Add apple.skip_codesign_simulator_bundles (#958)
This is a replacement for `apple.codesign_simulator_bundles` that uses
the newer `features` system. This has the benefit of being able to set
it on specific targets, and disable it for others. This way if you have
some targets that require signing, you can use:

```
features = ["-apple.skip_codesign_simulator_bundles"],
```

On those will still passing
`--features=apple.skip_codesign_simulator_bundles` if that works for
you.

Once the old option is removed there will be no filtering for when to
use this, so you can use it on whatever target types you want, and like
before it's up to you to manage the tradeoff between performance and
correctness.

Fixes https://github.com/bazelbuild/rules_apple/issues/710