Remove explicit --source-libraries argument to swift-stdlib-tool
swift-stdlib-tool automatically derives this path when it is not passed
based on the platform we provide. This way we can avoid this extra
`xcode-select` call and future proof this a bit more. This also makes it
easier to support custom toolchains, where the XcodeDefault.xctoolchain
part of this path logic was invalid.