Fix entitlement validation with appattest
f0404bc5e593f77d84820ad35fd016fc6a82eb24 introduced a regression where
the value of this entitlement was in the potential list, so it fell
through to the next conditional which validated they were equal, which
will never be the case. This also adds a test for this successful case.