Update min version for Swift dylib copying
To support the new back deployed concurrency libraries we need to run
swift-stdlib-tool for newer versions than before. This bumps the minimum
version until we can stop running this action to reflect this.
swift-stdlib-tool then does the right thing to only copy the concurrency
library, in the case you're deploying to a version newer than 12.2.

Part of #1253