Fix Swift dylib thinning on macOS with Xcode 13.2.1.
The current version of the script contained an assumption that all Swift dylibs that would get copied into a bundle would have the same architecture slices. This is no longer the case with back-deployed concurrency. On macOS, the core Swift libs only have x86_64 slices (because there is no need for arm64 slices since all macOS versions compatible with arm64 have Swift in the OS), but `libswift_Concurrency.dylib` has arm64(e) slices. This caused the script to break, attempting to call `lipo -thin` on an already-thin dylib.

Now, we keep track of a mapping of which dylibs had which slices so that we can intersect each one with the set that we want to keep and use that to drive the thinning logic.

PiperOrigin-RevId: 423172109