Make sure output directory for actool exists before invoking actool
Previously, each "actool" compile action declared an output directory
but didn't create it. Instead, it relied on Bazel to create the
directory before invoking the tool. With a non-remote spawn strategy,
Bazel would implicitly create it for you; but with the remote spawn
strategy, it is not created by the remote worker.

Additionally, this was also against what was documented for
declare_directory, which says that "You must create an action that
generates the directory".

This patches the "actool" wrapper to create an empty directory if one
doesn't exist at the specified output path, to make it works with remote
execution.

https://github.com/bazelbuild/rules_apple/pull/896

RELNOTES: None
PiperOrigin-RevId: 423797086