Ignore stderr from commands
On m1 machines xcrun and xcodebuild prints warnings like these:

```
objc[11792]: Class AMSupportURLConnectionDelegate is implemented in both /usr/lib/libauthinstall.dylib (0x207db1160) and /System/Library/PrivateFrameworks/MobileDevice.framework/Versions/A/MobileDevice (0x1186102b8). One of the two will be used. Which one is undefined.
```

I reported this 4/21 as FB9089778, but in the meantime we can just
ignore it.