Fix entitlement validation with appattest
https://github.com/bazelbuild/rules_apple/commit/e61b0d48e4a1ee271871178eb1994f236715d744 introduced a regression where
the value of this entitlement was in the potential list, so it fell
through to the next conditional which validated they were equal, which
will never be the case. This also adds a test for this successful case.

Based on PR #1308

PiperOrigin-RevId: 426183779
(cherry picked from commit 5587358e12f861d21359b7de887d04f3a774d402)