Rename a bunch of things to prepare for having two toolchains.
The long term goal being that by having two toolchains, they will eventually be
able to make use of exec_* attributes to mark one as Mac/darwin requiring. This
should avoid the complexity of having to individual annotate each thing on the toolchain.

This is a noop from a functionality pov.

- Rename `AppleSupportToolchainInfo` provider to `AppleSupportMacToolsToolchainInfo`
- Rename `_toolchain` attribute to `_mac_toolchain`
- Rename `apple_toolchain_info` arguments to `apple_mac_toolchain_info`
- Rename `apple_support_toolchain` rule to `apple_support_mac_tools_toolchain`
- Rename `toolchain_support` target to `mac_tools_toolchain`

RELNOTES: None
PiperOrigin-RevId: 441739808